chore: Move ViewModel And Model
디자인이 변경되면서 위치 변경
