+++ feature-main/workspace/src/main/java/com/seugi/workspace/feature/joinsuccess/JoinSuccessViewModel.kt
@@ -1,4 +1,4 @@
-package com.seugi.workspace.feature.selectingcode
+package com.seugi.workspace.feature.joinsuccess
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
@@ -6,7 +6,7 @@ import com.seugi.common.model.Result
 import com.seugi.common.utiles.DispatcherType
 import com.seugi.common.utiles.SeugiDispatcher
 import com.seugi.data.workspace.WorkspaceRepository
-import com.seugi.workspace.feature.selectingcode.model.SelectingCodeSideEffect
+import com.seugi.workspace.feature.joinsuccess.model.JoinSuccessSideEffect
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 import kotlinx.coroutines.CoroutineDispatcher
@@ -16,13 +16,13 @@ import kotlinx.coroutines.flow.receiveAsFlow
 import kotlinx.coroutines.launch
 
 @HiltViewModel
-class SelectingCodeViewModel @Inject constructor(
+class JoinSuccessViewModel @Inject constructor(
     private val workspaceRepository: WorkspaceRepository,
     @SeugiDispatcher(DispatcherType.IO) private val dispatcher: CoroutineDispatcher,
 ) : ViewModel() {
 
-    private val _selectingCodeSideEffect = Channel<SelectingCodeSideEffect>()
-    val selectingCodeSideEffect = _selectingCodeSideEffect.receiveAsFlow()
+    private val _joinSuccessSideEffect = Channel<JoinSuccessSideEffect>()
+    val joinSuccessSideEffect = _joinSuccessSideEffect.receiveAsFlow()
     fun workspaceApplication(role: String) {
         viewModelScope.launch(dispatcher) {
             workspaceRepository.workspaceApplication(
@@ -33,10 +33,10 @@ class SelectingCodeViewModel @Inject constructor(
             ).collectLatest {
                 when (it) {
                     is Result.Success -> {
-                        _selectingCodeSideEffect.send(SelectingCodeSideEffect.SuccessApplication)
+                        _joinSuccessSideEffect.send(JoinSuccessSideEffect.SuccessApplication)
                     }
                     is Result.Error -> {
-                        _selectingCodeSideEffect.send(SelectingCodeSideEffect.FiledApplication(it.throwable))
+                        _joinSuccessSideEffect.send(JoinSuccessSideEffect.FiledApplication(it.throwable))
                     }
                     is Result.Loading -> {
                     }
