+++ feature-main/workspace/src/main/java/com/seugi/workspace/feature/joinsuccess/model/JoinSuccessSideEffect.kt
@@ -1,0 +1,6 @@
+package com.seugi.workspace.feature.joinsuccess.model
+
+sealed class JoinSuccessSideEffect {
+    data object SuccessApplication : JoinSuccessSideEffect()
+    data class FiledApplication(val throwable: Throwable) : JoinSuccessSideEffect()
+}
