--- feature-main/workspace/src/main/java/com/seugi/workspace/feature/selectingcode/model/SelectingCodeSideEffect.kt
@@ -1,6 +1,0 @@
-package com.seugi.workspace.feature.selectingcode.model
-
-sealed class SelectingCodeSideEffect {
-    data object SuccessApplication : SelectingCodeSideEffect()
-    data class FiledApplication(val throwable: Throwable) : SelectingCodeSideEffect()
-}
